+++ src/lifish/components/LevelNumText.hpp
@@ -11,7 +11,7 @@ class LevelNumText : public Game::Component, public sf::Drawable {
 
 public:
 	explicit LevelNumText(Game::Entity& owner, 
-			const unsigned short levelnum);
+			const unsigned levelnum);
 
 	void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
 	void setOrigin(const sf::Vector2f& pos) override;
