+++ src/lifish/entities/Bomb.cpp
@@ -14,7 +14,7 @@ using Game::TILE_SIZE;
 using namespace Game::Conf::Bomb;
 
 Bomb::Bomb(const sf::Vector2f& pos, const Game::Player& source, 
-		const sf::Time& _fuseTime, const unsigned short _radius)
+		const sf::Time& _fuseTime, const unsigned _radius)
 	: Game::Entity(pos)
 	, fuseTime(_fuseTime)
 	, radius(_radius)
